+++ legend-engine-config/legend-engine-extensions-collection-generation/pom.xml
@@ -670,6 +670,11 @@
         <!-- SQL -->
         <dependency>
             <groupId>org.finos.legend.engine</groupId>
+            <artifactId>legend-engine-xt-sql-expression-pure</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.finos.legend.engine</groupId>
             <artifactId>legend-engine-xt-sql-expression-grammar</artifactId>
         </dependency>
         <dependency>
