+++ legend-engine-xts-sql/legend-engine-xt-sql-expression/legend-engine-xt-sql-expression-pure/src/main/resources/core_external_query_sql_expression.definition.json
@@ -4,6 +4,7 @@
   "dependencies": [
     "platform",
     "platform_dsl_store",
+    "core",
     "core_external_store_relational_postgres_sql_model"
   ]
 }
