+++ legend-engine-xts-sql/legend-engine-xt-sql-expression/legend-engine-xt-sql-expression-pure/src/main/resources/core_external_query_sql_expression/frameworkExtension.pure
@@ -3,7 +3,8 @@ import meta::external::query::sql::metamodel::*;
 import meta::pure::store::*;
 import meta::pure::router::routing::*;
 import meta::pure::extension::*;
-function meta::external::query::sql::transformation::queryToPure::tests::sqlExpressionFeatureExtension() : Extension[1]
+
+function meta::external::query::sql::expression::sqlExpressionFeatureExtension() : Extension[1]
 {
    let shared = ^Extension(
       type = 'sqlExpressionExtension',
