+++ legend-engine-xts-sql/legend-engine-xt-sql-expression/legend-engine-xt-sql-expression-tests-pure/src/main/resources/core_external_query_sql_expression_tests.definition.json
@@ -1,0 +1,11 @@
+{
+  "name": "core_external_query_sql_expression_tests",
+  "pattern": "(meta::external::query::sql::expression)(::.*)?",
+  "dependencies": [
+    "platform",
+    "core_functions_unclassified",
+    "core",
+    "core_external_extensions",
+    "core_external_compiler"
+  ]
+}
