+++ legend-engine-xts-sql/legend-engine-xt-sql-expression/legend-engine-xt-sql-expression-tests-pure/src/main/resources/core_external_query_sql_expression_tests/tests/query.legend
