+++ legend-engine-xts-sql/legend-engine-xt-sql-expression/legend-engine-xt-sql-expression-tests-pure/src/main/resources/core_external_query_sql_expression_tests/tests/testQuery.pure
@@ -12,9 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-function <<test.Test>> meta::external::query::sql::transformation::queryToPure::tests::testExpression():Boolean[1]
+function <<test.Test>> meta::external::query::sql::expression::tests::testExpression():Boolean[1]
 {
-  let res = meta::legend::compile(readFile('/core_external_query_sql/binding/fromPure/tests/expression/query.legend')->toOne());
+  let res = meta::legend::compile(readFile('/core_external_query_sql_expression_tests/tests/query.legend')->toOne());
   meta::pure::executionPlan::executionPlan(
     $res->filter(p|$p->instanceOf(FunctionDefinition))->toOne()->cast(@FunctionDefinition<Any>),
     meta::pure::extension::runtime::getExtensions(),
